feat: move menu and routes to src/config
