+++ scaffolds/admin-lite/.eslintrc
@@ -1,0 +1,60 @@
+{
+  "root": true,
+  "parser": "babel-eslint",
+  "extends": "eslint-config-airbnb",
+  "parserOptions": {
+    "ecmaVersion": 6,
+    "ecmaFeatures": {
+      "jsx": true,
+      "experimentalObjectRestSpread": true
+    }
+  },
+  "env": {
+    "browser": true,
+    "mocha": true
+  },
+  "plugins": ["vue", "babel"],
+  "rules": {
+    "react/prefer-stateless-function": 0,
+    "no-console": 0,
+    "jsx-a11y/label-has-for": 0,
+    "no-did-mount-set-state": 0,
+    "jsx-a11y/no-static-element-interactions": 0,
+    "jsx-a11y/anchor-has-content": 0,
+    "jsx-a11y/click-events-have-key-events": 0,
+    "jsx-a11y/anchor-is-valid": 0,
+    "react/no-array-index-key": 0,
+    "func-names": 0,
+    "arrow-body-style": 0,
+    "no-use-before-define": 0,
+    "react/sort-comp": 0,
+    "react/prop-types": 0,
+    "react/jsx-first-prop-new-line": 0,
+    "react/jsx-filename-extension": [
+      1,
+      {
+        "extensions": [".js", ".jsx"]
+      }
+    ],
+    "import/extensions": 0,
+    "import/no-unresolved": 0,
+    "import/no-extraneous-dependencies": 0,
+    "prefer-destructuring": 0,
+    "no-param-reassign": 0,
+    "no-return-assign": 0,
+    "max-len": 0,
+    "consistent-return": 0,
+    "no-redeclare": 0,
+    "react/require-extension": 0,
+    "react/no-danger": 0,
+    "react/forbid-prop-types": 0,
+    "comma-dangle": ["error", "always-multiline"],
+    "function-paren-newline": 0,
+    "object-curly-newline": 0,
+    "no-restricted-globals": 0,
+    "react/jsx-no-bind": 0,
+    "no-mixed-operators": 0,
+    "global-require": 0,
+    "linebreak-style": [0, "error", "windows"]
+  }
+}
