+++ scaffolds/admin-lite/src/layouts/HeaderAsideLayout/components/SideBar.vue
@@ -1,12 +1,12 @@
 <template>
   <scroll-bar>
     <div class="logo">
-      <img src="https://img.alicdn.com/tfs/TB13UQpnYGYBuNjy0FoXXciBFXa-242-134.png" width="40">
+      <img src="https://img.alicdn.com/tfs/TB13UQpnYGYBuNjy0FoXXciBFXa-242-134.png" width="40" />
       <span class="site-name">ADMIN LITE</span>
     </div>
     <el-menu
-      mode="vertical"
       :show-timeout="200"
+      mode="vertical"
       background-color="#00142a"
       text-color="hsla(0, 0%, 100%, .65)"
       active-text-color="#409EFF"
@@ -14,43 +14,42 @@
       <template v-for="item in asideMenuConfig">
         <router-link v-if="!item.children" :to="item.path" :key="item.name">
           <el-menu-item :index="item.path">
-            <i v-if="item.icon" :class="item.icon"></i>
-            <span v-if="item.name" slot="title">{{item.name}}</span>
+            <i v-if="item.icon" :class="item.icon" />
+            <span v-if="item.name" slot="title">{{ item.name }}</span>
           </el-menu-item>
         </router-link>
 
         <el-submenu v-else :index="item.name || item.path" :key="item.name">
           <template slot="title">
-            <i v-if="item && item.icon" :class="item.icon"></i>
-            <span v-if="item && item.name" slot="title">{{item.name}}</span>
+            <i v-if="item && item.icon" :class="item.icon" />
+            <span v-if="item && item.name" slot="title">{{ item.name }}</span>
           </template>
           <template v-for="child in item.children" v-if="!child.hidden">
             <router-link :to="item.path + child.path" :key="child.name">
               <el-menu-item :index="item.path + child.path">
-                <span v-if="child && child.name" slot="title">{{child.name}}</span>
+                <span v-if="child && child.name" slot="title">{{ child.name }}</span>
               </el-menu-item>
             </router-link>
           </template>
         </el-submenu>
-
       </template>
     </el-menu>
   </scroll-bar>
 </template>
 
 <script>
-import ScrollBar from './ScrollBar';
-import { asideMenuConfig } from '../../../menuConfig';
+import ScrollBar from "./ScrollBar";
+import { asideMenuConfig } from "@config/menu";
 
 export default {
   components: { ScrollBar },
-  name: 'SideBar',
+  name: "SideBar",
   props: {},
   data() {
     return {
-      asideMenuConfig,
+      asideMenuConfig
     };
-  },
+  }
 };
 </script>
 
@@ -72,7 +71,7 @@ export default {
   margin-left: 10px;
 }
 .sidebar-container {
-  box-shadow: 2px 0 6px rgba(0, 21, 41, .35);
+  box-shadow: 2px 0 6px rgba(0, 21, 41, 0.35);
   transition: width 0.28s;
   width: 256px !important;
   height: 100%;
